+++ NEWS.md
@@ -21,8 +21,12 @@ sudo make install
   - Add more columns to oauth_app ([#15015](https://github.com/CartoDB/cartodb/issues/15015))
   - Track OauthApp and OauthAppUser events in Segment ([#15055](https://github.com/CartoDB/cartodb/pull/15055))
   - Update Auth API swagger spec to include schemas and table_metadata grants ([#14998](https://github.com/CartoDB/cartodb/issues/14998))
+  - Allow developers to manage their OAuth apps in the dashboard ([#15031](https://github.com/CartoDB/cartodb/pull/15031))
+  - Scope for data_observatory_token ([#15089](https://github.com/CartoDB/cartodb/pull/15089))
 - Add number of employees and use case to user profile ([#14966](https://github.com/CartoDB/cartodb/pull/14966))
 - Add CARTO Data Source Request link ([CartoDB/product#441](https://github.com/CartoDB/product/issues/441))
+- Data Observatory token endpoint ([#15097](https://github.com/CartoDB/cartodb/pull/15097))
+- Add GET MFA status to EUMAPI ([CartoDB/cartodb#15101](https://github.com/CartoDB/cartodb/issues/15101))
 
 ### Bug fixes / enhancements
 - Fix API keys page when tables had certain reserved names ([#15059](https://github.com/CartoDB/cartodb/pull/15059))
@@ -33,7 +37,14 @@ sudo make install
 - Change embeds attribution character ([#14914](https://github.com/CartoDB/cartodb/issues/14914))
 - Fix disabled privacy button in Builder when there are no other public maps ([CartoDB/support#2163](https://github.com/CartoDB/support/issues/2163))
 - Include password confirmation in the delete mobile app modal ([CartoDB/support#2155](https://github.com/CartoDB/support/issues/2155))([#15061](https://github.com/CartoDB/cartodb/pull/15061))
+- Rename "Professional" Plan to "Individual" Plan ([#15069](https://github.com/CartoDB/cartodb/pull/15069))
 - The type of the tables_id column of user_tables has changed from integer to oid ([#15068](https://github.com/CartoDB/cartodb/issues/15068))
+- Revamp link to DB connectors feedback ([#1614](https://github.com/CartoDB/design/issues/1614))
+- Fix schema name in create API key permission ([#15082](https://github.com/CartoDB/cartodb/pull/15082))
+- Minor CSS fixes in Mobile Apps page ([#15090](https://github.com/CartoDB/cartodb/pull/15090))
+- Revert connectors link to previous version ([#15096](https://github.com/CartoDB/cartodb/pull/15096))
+- Fix broken link in oauth app page ([#15098](https://github.com/CartoDB/cartodb/issues/15098))
+- Include users API key for EUMAPI ([#15102](https://github.com/CartoDB/cartodb/issues/15102))
 
 4.29.0 (2019-07-15)
 -------------------